added iroh tests under thumbs.js
